pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "zodb-pgjsonb"
dynamic = ["version"]
description = "ZODB storage adapter using PostgreSQL JSONB with zodb-json-codec transcoding"
readme = "README.md"
license = "ZPL-2.1"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Zope Public License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database",
    "Framework :: ZODB",
]
dependencies = [
    "ZODB>=5.8",
    "zodb-json-codec",
    "psycopg[binary,pool]>=3.1",
    "zope.interface",
    "transaction",
    "ZConfig",
]

[project.urls]
Homepage = "https://github.com/bluedynamics/zodb-pgjsonb"
Repository = "https://github.com/bluedynamics/zodb-pgjsonb"
Issues = "https://github.com/bluedynamics/zodb-pgjsonb/issues"

[project.optional-dependencies]
s3 = ["zodb-s3blobs", "boto3>=1.26"]
test = [
    "pytest",
    "pytest-cov",
    "coverage[toml]>=7.0",
    "testcontainers[postgres]",
    "moto[s3]>=4.0",
    "zope.testing",
]
bench = [
    "relstorage>=4.0",
    "psycopg2-binary",
]

[tool.hatch.build.targets.sdist]
exclude = ["example/"]

[tool.hatch.build.targets.wheel]
packages = ["src/zodb_pgjsonb"]

[tool.hatch.version]
source = "vcs"
root = "../.."
raw-options = { local_scheme = "no-local-version" }

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
parallel = true
source = ["zodb_pgjsonb"]
relative_files = true

[tool.coverage.report]
show_missing = true
fail_under = 90
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "A",
    "B",
    "C4",
    "E",
    "F",
    "I",
    "RUF",
    "SIM",
    "T20",
    "UP",
    "W",
]
ignore = [
    "E501",
    "RUF001",
    "RUF002",
]

[tool.ruff.lint.isort]
force-single-line = true
from-first = true
lines-after-imports = 2
lines-between-types = 1
no-sections = true
order-by-type = false

[tool.ruff.lint.pycodestyle]
max-line-length = 120
max-doc-length = 120

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T201"]
"benchmarks/*" = ["T201"]
